examples/text_to_image/requirements.txt
accelerate>=0.16.0
torch==1.13.1 #https://github.com/huggingface/diffusers/issues/3565#issuecomment-1567984137
xformers==0.0.16
torchvision
transformers>=4.25.1
datasets
ftfy
tensorboard
Jinja2
